doc(doc, examples): add todos page
