+++ examples/webpack-two/project.config.json
@@ -7,7 +7,6 @@
         "minified": true,
         "newFeature": true
     },
-    "miniprogramRoot": "./dist/",
     "compileType": "miniprogram",
     "appid": "touristappid",
     "projectname": "github-tua-mp-webpack-two",
