+++ examples/webpack-two/src/app/app.json
@@ -1,6 +1,7 @@
 {
     "pages":[
         "pages/index/index",
+        "pages/todos/todos",
         "pages/logs/logs"
     ],
     "window":{
