+++ examples/webpack-two/src/pages/index/index.json
@@ -1,0 +1,3 @@
+{
+    "navigationBarTitleText": "tua-mp webpack two"
+}
