+++ examples/webpack-two/src/pages/index/index.wxml
@@ -1,108 +1,14 @@
-<view class="todoapp">
-    <view class="header">
-        <view class="h1">todos</view>
-        <input
-            focus
-            value="{{ newTodo }}"
-            class="new-todo"
-            bindinput="inputTodo"
-            bindconfirm="addTodo"
-            placeholder="What needs to be done?"
-        />
+<!--index.wxml-->
+<view class="container">
+    <view class="userinfo">
+        <button wx:if="{{!hasUserInfo && canIUse}}" open-type="getUserInfo" bindgetuserinfo="getUserInfo"> 获取头像昵称 </button>
+        <block wx:else>
+            <image bindtap="bindViewTap" class="userinfo-avatar" src="{{userInfo.avatarUrl}}" mode="cover"></image>
+            <text class="userinfo-nickname">{{userInfo.nickName}}</text>
+        </block>
     </view>
 
-    <view class="main" hidden="{{ !todos.length }}">
-        <checkbox-group bindchange="toggleAll">
-            <checkbox
-                id="toggle-all"
-                class="toggle-all {{ allDone ? 'checked' : '' }}"
-                value="{{ allDone }}"
-                checked="{{ allDone }}"
-            />
-            <label for="toggle-all">Mark all as complete</label>
-        </checkbox-group>
-
-        <view class="todo-list">
-            <view
-                class="li todo {{ todo.completed ? 'completed' : '' }} {{ todo.id === editedTodo.id ? 'editing' : '' }}"
-                wx:for="{{ filteredTodos }}"
-                wx:for-item="todo"
-                wx:for-index="index"
-                wx:key="id"
-            >
-                <view class="view">
-                    <checkbox-group data-idx="{{ index }}" bindchange="toggleTodo">
-                        <checkbox
-                            class="toggle {{ todo.completed ? 'checked' : '' }}"
-                            checked="{{ todo.completed }}"
-                        />
-                        <label data-idx="{{ index }}" bindlongpress="onPressTodo">
-                            {{ todo.title }}
-                        </label>
-                    </checkbox-group>
-                    <view
-                        class="destroy"
-                        data-idx="{{ index }}"
-                        bindtap="onTapRemove"
-                    />
-                </view>
-                <input
-                    class="edit"
-                    value="{{ todo.title }}"
-                    data-idx="{{ index }}"
-                    bindblur="onBlurTodo"
-                />
-            </view>
-        </view>
-    </view>
-
-    <view class="footer" hidden="{{ !todos.length }}">
-        <view class="todo-count">
-            {{ remaining }} {{ remainingStr }} left
-        </view>
-
-        <view class="filters">
-            <view class="li">
-                <view
-                    class="a {{ visibility == 'all' ? 'selected' : '' }}"
-                    data-filter="all"
-                    bindtap="changeFilter"
-                >
-                    All
-                </view>
-            </view>
-            <view class="li">
-                <view
-                    class="a {{ visibility == 'active' ? 'selected' : '' }}"
-                    data-filter="active"
-                    bindtap="changeFilter"
-                >
-                    Active
-                </view>
-            </view>
-            <view class="li">
-                <view
-                    class="a {{ visibility == 'completed' ? 'selected' : '' }}"
-                    data-filter="completed"
-                    bindtap="changeFilter"
-                >
-                    Completed
-                </view>
-            </view>
-        </view>
-
-        <view
-            class="clear-completed"
-            bindtap="removeCompleted"
-            style="visibility: {{ todos.length <= remaining ? 'hidden' : 'visible' }};"
-        >
-            Clear completed
-        </view>
-    </view>
-</view>
-
-<view class="info">
-    <view>Long-press to edit a todo</view>
-    <view><text style="font-weight: 400;">长按</text>编辑一个 todo</view>
-    <view class="a">Written by StEve Young\n(https://github.com/BuviewtStEve)</view>
+    <button class="usermotto" catchtap="gotoTodos">
+        <text class="user-motto">点我去 todos 页面</text>
+    </button>
 </view>
