+++ examples/webpack-two/src/pages/todos/todos.json
@@ -1,0 +1,3 @@
+{
+    "navigationBarTitleText": "tua-mp todos"
+}
