+++ examples/webpack-two/src/pages/todos/todos.wxml
@@ -1,0 +1,117 @@
+<view class="todoapp">
+    <view class="header">
+        <view class="h1">todos</view>
+        <input
+            focus
+            value="{{ newTodo }}"
+            class="new-todo"
+            bindinput="inputTodo"
+            bindconfirm="addTodo"
+            placeholder="What needs to be done?"
+        />
+    </view>
+
+    <view class="main" hidden="{{ !todos.length }}">
+        <checkbox-group bindchange="toggleAll">
+            <checkbox
+                id="toggle-all"
+                class="toggle-all {{ allDone ? 'checked' : '' }}"
+                value="{{ allDone }}"
+                checked="{{ allDone }}"
+            />
+            <label for="toggle-all">Mark all as complete</label>
+        </checkbox-group>
+
+        <view class="todo-list">
+            <view
+                class="li todo {{ todo.completed ? 'completed' : '' }} {{ todo.id === editedTodo.id ? 'editing' : '' }}"
+                wx:for="{{ filteredTodos }}"
+                wx:key="id"
+                wx:for-item="todo"
+                wx:for-index="index"
+            >
+                <view class="view">
+                    <checkbox-group
+                        data-index="{{ index }}"
+                        bindchange="toggleTodo"
+                    >
+                        <checkbox
+                            class="toggle {{ todo.completed ? 'checked' : '' }}"
+                            checked="{{ todo.completed }}"
+                        />
+                        <label
+                            data-todo="{{ todo }}"
+                            bindlongpress="onPressTodo"
+                        >
+                            {{ todo.title }}
+                        </label>
+                    </checkbox-group>
+
+                    <view
+                        class="destroy"
+                        data-index="{{ index }}"
+                        bindtap="onTapRemove"
+                    />
+                </view>
+
+                <input
+                    class="edit"
+                    value="{{ todo.title }}"
+                    data-index="{{ index }}"
+                    bindblur="onBlurTodo"
+                />
+            </view>
+        </view>
+    </view>
+
+    <view class="footer" hidden="{{ !todos.length }}">
+        <view class="todo-count">
+            {{ remaining }} {{ remainingStr }} left
+        </view>
+
+        <view class="filters">
+            <view class="li">
+                <view
+                    class="a {{ visibility == 'all' ? 'selected' : '' }}"
+                    data-filter="all"
+                    bindtap="changeFilter"
+                >
+                    All
+                </view>
+            </view>
+            <view class="li">
+                <view
+                    class="a {{ visibility == 'active' ? 'selected' : '' }}"
+                    data-filter="active"
+                    bindtap="changeFilter"
+                >
+                    Active
+                </view>
+            </view>
+            <view class="li">
+                <view
+                    class="a {{ visibility == 'completed' ? 'selected' : '' }}"
+                    data-filter="completed"
+                    bindtap="changeFilter"
+                >
+                    Completed
+                </view>
+            </view>
+        </view>
+
+        <view
+            class="clear-completed"
+            bindtap="removeCompleted"
+            style="{{ clearCompletedStyleStr }}"
+        >
+            Clear completed
+        </view>
+    </view>
+</view>
+
+<view class="info">
+    <view>Long-press to edit a todo</view>
+    <view><text style="font-weight: bold;">长按</text>编辑一个 todo</view>
+    <view class="a">Written by StEve Young</view>
+    <view class="a">Powered By tua-mp @github/@npm</view>
+</view>
